backend/requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
asgiref==3.3.0
astor==0.8.1
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
cached-property==1.5.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
django-pandas==0.6.4
django-templated-mail==1.1.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
djoser==2.1.0
drf-yasg==1.20.0
drfauto==1.5.2
et-xmlfile==1.0.1
gast==0.2.2
google-pasta==0.2.0
grpcio==1.36.1
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
joblib==1.0.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
loguru==0.5.3
Markdown==3.3.4
MarkupSafe==1.1.1
multidict==5.1.0
numpy==1.19.0
oauthlib==3.1.0
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==20.4
pandas==1.2.3
protobuf==3.15.6
psycopg2==2.8.6
pycparser==2.20
pydantic==1.8.1
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2019.3
PyYAML==5.4.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sklearn==0.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.2.1
sqlparse==0.4.1
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.11
websocket-client==0.56.0
Werkzeug==1.0.1
win32-setctime==1.0.3
wrapt==1.12.1
xlrd==2.0.1
yarl==1.6.3
zipp==3.4.0
